+++ glmat33/block.go
@@ -11,7 +11,7 @@ func (v Vec4) V2() Vec2 {
 }
 
 // Return a Vec4 constructed using elements of v appended by w.
-func (v Vec3) V4(w float32) Vec4 {
+func (v Vec3) V4(w float64) Vec4 {
 	return Vec4{v[0], v[1], v[2], w}
 }
 
@@ -21,12 +21,12 @@ func (v Vec3) V2() Vec2 {
 }
 
 // Return a Vec4 constructed using elements of v appended by z and w.
-func (v Vec2) V4(z, w float32) Vec4 {
+func (v Vec2) V4(z, w float64) Vec4 {
 	return Vec4{v[0], v[1], z, w}
 }
 
 // Return a Vec3 constructed using elements of v appended by z.
-func (v Vec2) V3(z float32) Vec3 {
+func (v Vec2) V3(z float64) Vec3 {
 	return Vec3{v[0], v[1], z}
 }
