+++ glmat42/scale.go
@@ -1,28 +1,28 @@
 package glmat42
 
 // lambda * v.
-func (v Vec4) Scaled(lambda float32) (scaled Vec4) {
+func (v Vec4) Scaled(lambda float64) (scaled Vec4) {
 	scaled = v
 	scaled.Scale(lambda)
 	return scaled
 }
 
 // lambda * v.
-func (v Vec3) Scaled(lambda float32) (scaled Vec3) {
+func (v Vec3) Scaled(lambda float64) (scaled Vec3) {
 	scaled = v
 	scaled.Scale(lambda)
 	return scaled
 }
 
 // lambda * v.
-func (v Vec2) Scaled(lambda float32) (scaled Vec2) {
+func (v Vec2) Scaled(lambda float64) (scaled Vec2) {
 	scaled = v
 	scaled.Scale(lambda)
 	return scaled
 }
 
 // v *= lambda.
-func (v *Vec4) Scale(lambda float32) *Vec4 {
+func (v *Vec4) Scale(lambda float64) *Vec4 {
 	v[0] *= lambda
 	v[1] *= lambda
 	v[2] *= lambda
@@ -32,7 +32,7 @@ func (v *Vec4) Scale(lambda float32) *Vec4 {
 }
 
 // v *= lambda.
-func (v *Vec3) Scale(lambda float32) *Vec3 {
+func (v *Vec3) Scale(lambda float64) *Vec3 {
 	v[0] *= lambda
 	v[1] *= lambda
 	v[2] *= lambda
@@ -41,7 +41,7 @@ func (v *Vec3) Scale(lambda float32) *Vec3 {
 }
 
 // v *= lambda.
-func (v *Vec2) Scale(lambda float32) *Vec2 {
+func (v *Vec2) Scale(lambda float64) *Vec2 {
 	v[0] *= lambda
 	v[1] *= lambda
